pyproject.toml
[tool.poetry]
name = "autotrade"
version = "0.1.0"
description = ""
authors = ["Juan David Rengifo Castro <jdrengifoc@eafit.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
polygon-api-client = "^1.14.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
